requirements.txt
gmusicapi==13.0.0
spotipy==2.13.0
python-dotenv==0.14.0
